refactor: util 패키지로 관련 클래스들 이동
