+++ src/main/java/lotto/util/Validator.java
@@ -1,8 +1,10 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import lotto.domain.Lotto;
+import lotto.domain.LottoNumber;
 
 public class Validator {
 
@@ -13,11 +15,13 @@ public class Validator {
             throw new IllegalArgumentException("입력된 수가 정수가 아닙니다.");
         }
     }
+
     public static void isNegative(String input) {
         if(Integer.parseInt(input) < 1) {
             throw new IllegalArgumentException("음수가 입력되었습니다.");
         }
     }
+
     public static void checkSizeOfLotto(List<LottoNumber> lottoNumbers) {
         if (lottoNumbers.size() != Lotto.LOTTO_NUMBER_SIZE) {
             throw new IllegalArgumentException("로또 번호는 6개이어야 합니다.");
